# docs/API_Spec.yaml
openapi: 3.0.3
info:
  title: Vibe Money Book API
  version: 1.0.0
  description: 語音記帳應用 API — 透過自然語言實現一句話記帳
  contact:
    name: Development Team
servers:
  - url: http://localhost:3000/api/v1
    description: Development Server
...
paths:
...
  /budget/summary:
    get:
      tags: [Budget]
      summary: 取得本月預算摘要
      operationId: getBudgetSummary
      security:
        - BearerAuth: []
      parameters:
        - name: month
          in: query
          description: 月份 (YYYY-MM)，預設當月
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
            example: '2026-03'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    $ref: '#/components/schemas/BudgetSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /budget/categories:
    get:
      tags: [Budget]
      summary: 取得類別預算列表
      operationId: getCategoryBudgets
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          description: 篩選類別類型（income 或 expense）
          schema:
            type: string
            enum: [income, expense]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      categories:
                        type: array
                        items:
                          $ref: '#/components/schemas/CategoryBudget'

    post:
      tags: [Budget]
      summary: 新增自訂類別（PRD-F-012）
      description: 使用者確認 AI 建議的新類別時呼叫此端點。類別數量上限 20。
      operationId: createCategory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category]
              properties:
                category:
                  type: string
                  minLength: 2
                  maxLength: 50
                  example: 寵物
                type:
                  type: string
                  enum: [income, expense]
                  default: expense
                  description: 類別類型（收入或支出）
                budget_limit:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
      responses:
        '201':
          description: 類別新增成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: 類別新增成功
                  data:
                    $ref: '#/components/schemas/CategoryBudget'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: 類別已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 類別數量已達上限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [Budget]
      summary: 批量更新類別預算
      operationId: updateCategoryBudgets
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [categories]
              properties:
                categories:
                  type: array
                  items:
                    type: object
                    required: [category, budget_limit]
                    properties:
                      category:
                        type: string
                      budget_limit:
                        type: number
                        format: decimal
                        minimum: 0
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      categories:
                        type: array
                        items:
                          $ref: '#/components/schemas/CategoryBudget'
        '400':
          $ref: '#/components/responses/BadRequest'
  /budget/categories/{category}:
    delete:
      tags: [Budget]
      summary: 刪除自訂類別（PRD-F-012）
      description: 刪除使用者自訂類別。預設類別不可刪除。關聯交易記錄將重新歸類至「其他」。
      operationId: deleteCategory
      security:
        - BearerAuth: []
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 刪除成功
        '400':
          description: 不可刪除預設類別
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BudgetSummary:
      type: object
      properties:
        month:
          type: string
        monthly_budget:
          type: number
        total_spent:
          type: number
        remaining:
          type: number
        used_ratio:
          type: number
        total_income:
          type: number
          description: 當月總收入
        total_asset:
          type: number
          description: 當月總資產（收入 - 支出）
        transaction_count:
          type: integer
        category_summary:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              spent:
                type: number
              budget_limit:
                type: number
              used_ratio:
                type: number
    CategoryBudget:
      type: object
      properties:
        category:
          type: string
        type:
          type: string
          enum: [income, expense]
          default: expense
          description: 類別類型（收入或支出）
        budget_limit:
          type: number
          format: decimal
        is_custom:
          type: boolean
          description: 是否為使用者自訂類別
          default: false
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              reason:
                type: string
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: 參數驗證失敗
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未認證或 Token 過期
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 資源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...